requirements.txt
fastapi
google-cloud-storage
uvicorn[standard]
pandas
pyarrow